requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
cryptography
PyJWT
requests
